refactor: 테스트 리포매팅
