--- src/test/java/kitchenpos/fixture/TableGroupFixture.java
@@ -1,14 +1,0 @@
-package kitchenpos.fixture;
-
-import java.util.List;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
-
-public class TableGroupFixture {
-
-    public static TableGroup createTableGroup(final OrderTable... orderTables) {
-        final TableGroup tableGroup = new TableGroup();
-        tableGroup.setOrderTables(List.of(orderTables));
-        return tableGroup;
-    }
-}
